Update exercise Grid view layout

+++ Data/DTOs/FoodJournal/CreateFoodItemDto.cs
@@ -8,11 +8,11 @@ public class CreateFoodItemDto
 
     public string Units { get; set; } = "Each";
 
-    public int Calories { get; set; }
+    public double Calories { get; set; }
 
-    public int Carbs { get; set; }
+    public double Carbs { get; set; }
 
-    public int Fats { get; set; }
+    public double Fats { get; set; }
 
-    public int Proteins { get; set; }
+    public double Proteins { get; set; }
 }
